test(server): stabilize roots-discovery integration flake
Serialize test files (fileParallelism: false). The real-server
mcp-integration tests drive an MCP roots-discovery listRoots SSE round
trip on a 1s budget; running ~75 files across parallel workers starved
the server event loop, so the round trip lost the race and discovery fell
back to global — a flaky failure unrelated to product behavior.

Also fold the new http-hardening real-server tests into mcp-integration
(as a trailing describe) so the two heavy server bootstraps run
sequentially and never coexist. No product code changes.
